feat(game): WebView 게임 4개 추가 + HTML 기준 이름·이모지 동기화
- assets/games/: 탭배틀·퀴즈왕·카드매칭·반응속도 standalone HTML 추가
- webview_game_page.dart: GameChannel JS채널로 점수 수신 → earn() 적립
- game_select_page.dart: HTML 기준 8개 게임 (Flame4 + WebView4) 재작성
  - 이름 정정: 무한점프→러너, 낚시대전→코인받기, 보상금액 HTML 기준 일치
- global_rank_page.dart: 탭 8개 + 가로 스크롤로 전환
- backend/games.js: tap/quiz/memory/reflex ID 및 보상 공식 추가
- IA_document.html: 게임 스펙 섹션 8개 게임으로 업데이트, webview_flutter 패키지 추가
- pubspec.yaml: webview_flutter ^4.7.0 + assets/games/ 등록

+++ lib/pages/game/games/catch_game_page.dart
@@ -66,7 +66,7 @@ class CatchGame extends FlameGame with PanDetector, TapCallbacks {
       final iRect = Rect.fromCircle(center: Offset(item.position.x, item.position.y), radius: itemR);
       final bRect = Rect.fromLTWH(basket.position.x, basket.position.y, basketW, basketH);
       if (iRect.overlaps(bRect)) {
-        score += item.type == 2 ? -10 : 10 + item.type * 5;
+        score += item.type == 2 ? -10 : (10 + item.type * 5).toInt();
         item.removeFromParent(); _items.remove(item);
         continue;
       }
